requirements.txt
torchvision
torchaudio
numpy
pandas
scikit-learn
tqdm
questionary